+++ sample/src/main/java/gr/aueb/mscis/sample/resource/ProgramResource.java
@@ -17,11 +17,15 @@ import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.UriBuilder;
 import javax.ws.rs.core.UriInfo;
+
+import gr.aueb.mscis.sample.model.Epopteia;
 import gr.aueb.mscis.sample.model.Epoptis;
 import gr.aueb.mscis.sample.model.Program;
+import gr.aueb.mscis.sample.service.EpopteiaService;
 import gr.aueb.mscis.sample.service.EpoptisService;
 import gr.aueb.mscis.sample.service.ProgramService;
 
@@ -37,74 +41,118 @@ public class ProgramResource extends AbstractResource {
 	@Context 
 	UriInfo uriInfo;
 
+	//lista me ola ta programata
 	@GET 
 	@Produces(MediaType.APPLICATION_JSON)
-	public List<ProgramInfo> listAllPrograms() {
+	public List<ProgramInfo> listAllPrograms(@QueryParam("username") String user,@QueryParam("password") String pass) {
 		EntityManager em = getEntityManager();
-		ProgramService programService = new ProgramService(em);
-		List<Program> programs = programService.findAllPrograms();
-
-		List<ProgramInfo> programInfo = ProgramInfo.wrap(programs);
-
+		EpoptisService es = new EpoptisService(em);
+		boolean access = es.logIn(user, pass);
+		if(access) {
+			ProgramService programService = new ProgramService(em);
+			List<Program> programs = programService.findAllPrograms();
+	
+			List<ProgramInfo> programInfo = ProgramInfo.wrap(programs);
+			em.close();
+			return programInfo;
+		}
 		em.close();
+		List<ProgramInfo> programInfo = new ArrayList<ProgramInfo>();
 		return programInfo;
 
 	}
 	
+	//leptomeries gia ena programa me sugkekrimeno id
 	@GET
 	@Path("{programId:[0-9]*}")
 	@Produces(MediaType.APPLICATION_JSON)
-	public ProgramInfo getProgramDetails(@PathParam("programId") int programId) {
+	public ProgramInfo getProgramDetails(@PathParam("programId") int programId,@QueryParam("username") String user,@QueryParam("password") String pass) {
 
 		EntityManager em = getEntityManager();
-
-		ProgramService programService = new ProgramService(em);
-		Program program = programService.findProgramById(programId);
-
-		ProgramInfo programInfo = ProgramInfo.wrap(program);
-		em.close();
-
-		return programInfo;
-
-	} 
-	
-	@GET
-	@Path("{programId:[0-9]*}/anaforaEpoptwn")
-	@Produces(MediaType.APPLICATION_JSON)
-	public List<AnaforaEpoptwnInfo> getAnaforaEpoptwn(@PathParam("programId") int programId) {
-
-		EntityManager em = getEntityManager();
-		
-		List<AnaforaEpoptwnInfo> stoixeia = new ArrayList();
-		
 		EpoptisService es = new EpoptisService(em);
-		List<Epoptis> epoptes = es.findAllEpoptes();
-		for(Epoptis epoptis : epoptes) {
-			stoixeia.add(new AnaforaEpoptwnInfo(epoptis,programId));
+		boolean access = es.logIn(user, pass);
+		if(access) {
+			ProgramService programService = new ProgramService(em);
+			Program program = programService.findProgramById(programId);
+	
+			ProgramInfo programInfo = ProgramInfo.wrap(program);
+			em.close();
+	
+			return programInfo;
 		}
 		em.close();
 
-		return stoixeia;
-
+		return new ProgramInfo();
 	} 
 	
+	//paragogh anaforas epoptwn me tis epopties tou ka8enos
+	@GET
+	@Path("{programId:[0-9]*}/anaforaEpoptwn")
+	@Produces(MediaType.APPLICATION_JSON)
+	public List<AnaforaEpoptwnInfo> getAnaforaEpoptwn(@PathParam("programId") int programId,@QueryParam("username") String user,@QueryParam("password") String pass) {
+		EntityManager em = getEntityManager();
+		EpoptisService es = new EpoptisService(em);
+		List<AnaforaEpoptwnInfo> stoixeia = new ArrayList();
+		boolean access = es.logIn(user, pass);
+		if(access) {
+			List<Epoptis> epoptes = es.findAllEpoptes();
+			for(Epoptis epoptis : epoptes) {
+				stoixeia.add(new AnaforaEpoptwnInfo(epoptis,programId));
+			}
+			em.close();
+			return stoixeia;
+		}
+		
+		em.close();
+		return stoixeia;
+	}
+	
+	//paragogh anaforas epoptiwn me tous epoptes kathemias
+	@GET
+	@Path("{programId:[0-9]*}/anaforaEpoptiwn")
+	@Produces(MediaType.APPLICATION_JSON)
+	public List<AnaforaEpoptiwnInfo> getAnaforaEpoptiwn(@PathParam("programId") int programId,@QueryParam("username") String user,@QueryParam("password") String pass) {
+		EntityManager em = getEntityManager();
+		EpoptisService es = new EpoptisService(em);
+		List<AnaforaEpoptiwnInfo> stoixeia = new ArrayList();
+		boolean access = es.logIn(user, pass);
+		if(access) {
+			EpopteiaService eps = new EpopteiaService(em);
+			List<Epopteia> epopteies = eps.findAllEpopteies();
+			for(Epopteia epopteia : epopteies) {
+				if(epopteia.getProgram().getId()==programId)
+					stoixeia.add(new AnaforaEpoptiwnInfo(epopteia));
+			}
+			em.close();
+			return stoixeia;
+		}
+		em.close();
+		return stoixeia;
+	} 
+	
+	//dhmiourgia neou programmatos
 	@POST
 	@Consumes(MediaType.APPLICATION_JSON)
-	public Response createProgram(ProgramInfo programInfo) {
+	public Response createProgram(ProgramInfo programInfo,@QueryParam("username") String user,@QueryParam("password") String pass) {
 
 		EntityManager em = getEntityManager();
-
-		Program program = programInfo.getProgram(em);
-
-		ProgramService programService = new ProgramService(em);
-		program = programService.saveProgram(program);
-
-		UriBuilder ub = uriInfo.getAbsolutePathBuilder();
-		URI newProgramUri = ub.path(Integer.toString(program.getId())).build();
-
+		EpoptisService es = new EpoptisService(em);
+		boolean access = es.logIn(user, pass);
+		if(access) {
+			Program program = programInfo.getProgram(em);
+	
+			ProgramService programService = new ProgramService(em);
+			program = programService.saveProgram(program);
+	
+			UriBuilder ub = uriInfo.getAbsolutePathBuilder();
+			URI newProgramUri = ub.path(Integer.toString(program.getId())).build();
+	
+			em.close();
+	
+			return Response.created(newProgramUri).build();
+		}
 		em.close();
-
-		return Response.created(newProgramUri).build();
+		return Response.status(Status.UNAUTHORIZED).build();
 	}
 	
 //	@PUT
@@ -124,26 +172,31 @@ public class ProgramResource extends AbstractResource {
 //		return Response.ok().build();
 //	}
  
+	//diagrafi enos programmatos me vash to id
 	@DELETE
 	@Path("{programId:[0-9]*}")
-	public Response deleteProgram(@PathParam("programId") int programId) {
+	public Response deleteProgram(@PathParam("programId") int programId,@QueryParam("username") String user,@QueryParam("password") String pass) {
 
 		EntityManager em = getEntityManager();
-		
-		ProgramService programService = new ProgramService(em);
-		
-		Program program = programService.findProgramById(programId);
-
-		boolean result = programService.deleteProgram(program);
-		
-		if (!result) {
+		EpoptisService es = new EpoptisService(em);
+		boolean access = es.logIn(user, pass);
+		if(access) {
+			ProgramService programService = new ProgramService(em);
+			
+			Program program = programService.findProgramById(programId);
+	
+			boolean result = programService.deleteProgram(program);
+			
+			if (!result) {
+				em.close();
+				return Response.status(Status.NOT_FOUND).build();
+			}
+	
 			em.close();
-			return Response.status(Status.NOT_FOUND).build();
+			return Response.ok().build();
 		}
-
 		em.close();
-		return Response.ok().build();
-
+		return Response.status(Status.UNAUTHORIZED).build();
 	}
